Add Icinga service for passive checks
The configuration and jobs were added to allow Icinga monitoring for
specific high value Jenkins jobs. While the Jenkins side works
correctly, Icinga unfortunately did not know what to do with the passive
checks. I believe this is because the service is not defined anywhere.

This commit adds in the defined passive checks on Icinga so that it can
successfully receive the service.