Require a lower level class in govuk_postgresql::extension
We would like to be able to use govuk_postgresql::extension in combination
with postgresql::server::database, in which case Postgresql::Server::Db[$db]
will not be defined. Changing the `require` to
Postgresql::Server::Database[$db] allows us to use this defined type in that
context, and it continues to work in the original use case too.